Refactor resource template package.
