+++ resource/template/resource.go
@@ -5,10 +5,8 @@ package template
 
 import (
 	"bytes"
-	"crypto/md5"
 	"errors"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"os"
 	"os/exec"
@@ -16,7 +14,6 @@ import (
 	"path/filepath"
 	"strconv"
 	"strings"
-	"syscall"
 	"text/template"
 
 	"github.com/BurntSushi/toml"
@@ -25,11 +22,6 @@ import (
 )
 
 var replacer = strings.NewReplacer("/", "_")
-
-// StoreClient is used to swap out the
-type StoreClient interface {
-	GetValues(keys []string) (map[string]interface{}, error)
-}
 
 // TemplateResourceConfig holds the parsed template resource.
 type TemplateResourceConfig struct {
@@ -78,35 +70,8 @@ func (t *TemplateResource) setVars() error {
 	if err != nil {
 		return err
 	}
-	t.Vars = cleanKeys(vars)
+	t.Vars = cleanKeys(vars, config.Prefix())
 	return nil
-}
-
-func appendPrefix(prefix string, keys []string) []string {
-	s := make([]string, len(keys))
-	for i, k := range keys {
-		s[i] = path.Join(prefix, k)
-	}
-	return s
-}
-
-// cleanKeys is used to transform the path based keys we
-// get from the StoreClient to a more friendly format.
-func cleanKeys(vars map[string]interface{}) map[string]interface{} {
-	clean := make(map[string]interface{}, len(vars))
-	prefix := config.Prefix()
-	for key, val := range vars {
-		clean[pathToKey(key, prefix)] = val
-	}
-	return clean
-}
-
-// pathToKey translates etcd key paths into something more suitable for use
-// in Golang templates. Turn /prefix/key/subkey into key_subkey.
-func pathToKey(key, prefix string) string {
-	key = strings.TrimPrefix(key, prefix)
-	key = strings.TrimPrefix(key, "/")
-	return replacer.Replace(key)
 }
 
 // createStageFile stages the src configuration file by processing the src
@@ -295,67 +260,4 @@ func ProcessTemplateResources(s StoreClient) []error {
 		log.Debug("Processing of template resource " + p + " complete")
 	}
 	return runErrors
-}
-
-// fileStat return a fileInfo describing the named file.
-func fileStat(name string) (fi fileInfo, err error) {
-	if isFileExist(name) {
-		f, err := os.Open(name)
-		defer f.Close()
-		if err != nil {
-			return fi, err
-		}
-		stats, _ := f.Stat()
-		fi.Uid = stats.Sys().(*syscall.Stat_t).Uid
-		fi.Gid = stats.Sys().(*syscall.Stat_t).Gid
-		fi.Mode = stats.Sys().(*syscall.Stat_t).Mode
-		h := md5.New()
-		io.Copy(h, f)
-		fi.Md5 = fmt.Sprintf("%x", h.Sum(nil))
-		return fi, nil
-	} else {
-		return fi, errors.New("File not found")
-	}
-}
-
-// sameConfig reports whether src and dest config files are equal.
-// Two config files are equal when they have the same file contents and
-// Unix permissions. The owner, group, and mode must match.
-// It return false in other cases.
-func sameConfig(src, dest string) (bool, error) {
-	if !isFileExist(dest) {
-		return false, nil
-	}
-	d, err := fileStat(dest)
-	if err != nil {
-		return false, err
-	}
-	s, err := fileStat(src)
-	if err != nil {
-		return false, err
-	}
-	if d.Uid != s.Uid {
-		log.Info(fmt.Sprintf("%s has UID %d should be %d", dest, d.Uid, s.Uid))
-	}
-	if d.Gid != s.Gid {
-		log.Info(fmt.Sprintf("%s has GID %d should be %d", dest, d.Gid, s.Gid))
-	}
-	if d.Mode != s.Mode {
-		log.Info(fmt.Sprintf("%s has mode %s should be %s", dest, os.FileMode(d.Mode), os.FileMode(s.Mode)))
-	}
-	if d.Md5 != s.Md5 {
-		log.Info(fmt.Sprintf("%s has md5sum %s should be %s", dest, d.Md5, s.Md5))
-	}
-	if d.Uid != s.Uid || d.Gid != s.Gid || d.Mode != s.Mode || d.Md5 != s.Md5 {
-		return false, nil
-	}
-	return true, nil
-}
-
-// isFileExist reports whether path exits.
-func isFileExist(fpath string) bool {
-	if _, err := os.Stat(fpath); os.IsNotExist(err) {
-		return false
-	}
-	return true
 }
