+++ resource/template/store_client.go
@@ -1,0 +1,6 @@
+package template
+
+// StoreClient is used to swap out the backend store.
+type StoreClient interface {
+	GetValues(keys []string) (map[string]interface{}, error)
+}
